+++ albums.json
@@ -1,1 +1,1 @@
-[{"date":"1990-01-01","spotify":true}]
+[{"date":"1990-01-01","spotify":true},{"date":"2018-01-01","spotify":false}]
