+++ genre.json
@@ -1,1 +1,1 @@
-[{"name":"Rock"}]
+[{"name":"Rock"},{"name":"Comedy"}]
