+++ main.rb
@@ -5,4 +5,10 @@ def main
   app.run
 end
 
+def welcome
+  puts 'Welcome to your Catalog of things!'
+  puts ' '
+end
+
+welcome
 main
